Split command categories into separate files
Further separate commands so that each category of command is in it's
own file. The goal is to keep the number of callbacks defined for each
command to a reasonable amount per file.
