+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+export{ CommandEntry } from './commandEntry';
+export{ accessCommands } from './access';
+export{ textCommands   } from './text';
+export{ hubCommands    } from './hub';
+export{ navCommands    } from './nav';
